requirements.txt
Flask==3.1.2
python-dotenv==1.1.1
supabase==2.18.1